+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.88.0] - 2020-02-11
 ### Changed
 - Sort `package.json` dependencies when adding with `vtex setup`.
 - Use `open` package instead of `opn`.